2023-autumn-ab-python-hw-5-restful-api-redfox193/requirements.txt
fastapi
uvicorn
pydantic_settings
pytest
pytest-cov
httpx